bugs fixed

+++ src/main/java/school/mjc/stage0/module3/task4/FormulaCreator.java
@@ -4,9 +4,9 @@ public class FormulaCreator {
     public static void main(String[] args) {
         int a, b;
         a = b = 0;
-        double ans;
+        int ans;
 
-        ans = (9 * a * a - 5 * b + 2 + a - 7) * (a + b - 4 * a * b) / 2.0;
+        ans = (9 * a * a - 5 * b + 2 + a - 7) * ((a + b - 4 * a * b) / 2);
 
 
     }
